+++ components/analytics/metric/noop/src/main/java/com/flipperdevices/metric/noop/NoopMetricApiImpl.kt
@@ -3,6 +3,7 @@ package com.flipperdevices.metric.noop
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.metric.api.MetricApi
 import com.flipperdevices.metric.api.events.ComplexEvent
+import com.flipperdevices.metric.api.events.SessionState
 import com.flipperdevices.metric.api.events.SimpleEvent
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
@@ -12,4 +13,6 @@ class NoopMetricApiImpl @Inject constructor() : MetricApi {
     override fun reportSimpleEvent(simpleEvent: SimpleEvent, arg: String?) = Unit
 
     override fun reportComplexEvent(complexEvent: ComplexEvent) = Unit
+
+    override fun reportSessionState(state: SessionState) = Unit
 }